+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.3.7](https://github.com/Gangio-App/for-desktop/compare/v1.3.6...v1.3.7) (2026-04-04)
+
+### Bug Fixes
+
+* add 5-second safety timeout to splash screen to prevent hangs if update check doesn't respond.
+* wrap auto-updater checks in try-catch to prevent startup crashes.
+
 ## [1.3.6](https://github.com/Gangio-App/for-desktop/compare/v1.3.5...v1.3.6) (2026-04-04)
 
 ### Bug Fixes